+++ voxygen/src/anim/bird_medium/mod.rs
@@ -103,7 +103,9 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Chicken, _) => (4.0, 3.0),
                 (Goose, _) => (5.0, 5.0),
                 (Peacock, _) => (4.0, 7.0),
-                (Eagle, _) => (3.5, 5.0),
+                (Eagle, _) => (2.5, 5.0),
+                (Snowyowl, _) => (2.5, 5.0),
+                (Parrot, _) => (0.5, 4.5),
             },
             chest: match (body.species, body.body_type) {
                 (Duck, _) => (0.0, 5.0),
@@ -111,6 +113,8 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Goose, _) => (0.0, 8.0),
                 (Peacock, _) => (0.0, 10.0),
                 (Eagle, _) => (0.0, 8.0),
+                (Snowyowl, _) => (0.0, 4.5),
+                (Parrot, _) => (0.0, 5.0),
             },
             tail: match (body.species, body.body_type) {
                 (Duck, _) => (-3.0, 1.5),
@@ -118,6 +122,8 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Goose, _) => (-5.0, 3.0),
                 (Peacock, _) => (-5.5, 2.0),
                 (Eagle, _) => (-8.0, -4.0),
+                (Snowyowl, _) => (-6.0, -2.0),
+                (Parrot, _) => (-8.0, -2.0),
             },
             wing: match (body.species, body.body_type) {
                 (Duck, _) => (2.75, 0.0, 1.0),
@@ -125,6 +131,8 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Goose, _) => (3.75, -1.0, 2.0),
                 (Peacock, _) => (3.0, 0.0, 1.0),
                 (Eagle, _) => (3.0, -8.0, 4.0),
+                (Snowyowl, _) => (3.5, -5.5, 4.0),
+                (Parrot, _) => (2.0, -4.5, 3.0),
             },
             foot: match (body.species, body.body_type) {
                 (Duck, _) => (2.0, -1.5, 4.0),
@@ -132,6 +140,8 @@ impl<'a> From<&'a comp::bird_medium::Body> for SkeletonAttr {
                 (Goose, _) => (2.0, -1.5, 7.0),
                 (Peacock, _) => (2.0, -2.5, 8.0),
                 (Eagle, _) => (2.0, -2.0, 8.0),
+                (Snowyowl, _) => (1.5, -2.5, 7.0),
+                (Parrot, _) => (1.5, -3.0, 3.0),
             },
         }
     }